requirements.txt
gitpython
fastapi
sentence-transformers
chromadb
openai
python-dotenv
pandas
uvicorn
docker